pyproject.toml
[project]
name = "dev-utils"
version = "0.1.0"
description = ""
authors = [{ name = "karthik", email = "karthikkannan@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "invoke>=1.7.1",
    "pyjwt[crypto]>=2.10.1",
]
